update exercise and homefragment

+++ app/src/main/java/com/example/healthtracking/CardView/Food.java
@@ -3,17 +3,17 @@ package com.example.healthtracking.CardView;
 import android.widget.ProgressBar;
 
 public class Food {
-    private int calories;
+    private double calories;
     private int lit;
     private ProgressBar progressBar;
 
-    public Food(int calories, int lit, ProgressBar progressBar){
+    public Food(double calories, int lit, ProgressBar progressBar){
         this.calories = calories;
         this.lit = lit;
         this.progressBar = progressBar;
     }
 
-    public int getCalories() {
+    public double getCalories() {
         return calories;
     }
